changed double to decimal in TroopLevel

+++ DumDum/DumDum/Models/Entities/TroopLevel.cs
@@ -9,10 +9,10 @@ namespace DumDum.Models.Entities
         [Key]
         public int TroopLevelId { get; set; }
         public int Level { get; set; }
-        public double Attack { get; set; }
-        public double Defence { get; set; }
-        public double Speed { get; set; }
-        public double Consumption { get; set; }
+        public decimal Attack { get; set; }
+        public decimal Defence { get; set; }
+        public decimal Speed { get; set; }
+        public decimal Consumption { get; set; }
         public int Cost { get; set; }
         public int ConstTime { get; set; }
         public int TroopTypeId { get; set; }
